requirements.txt
# Web framework
flask

# Machine learning / NLP
torch
transformers
openai

# Data sources / APIs
yfinance

# Environment configuration
python-dotenv

# Database ORM
peewee

# Data analysis / Technical indicators
pandas_ta